Move PassField to own package
