+++ app/src/main/java/nz/eloque/foss_wallet/model/field/PassField.kt
@@ -1,0 +1,31 @@
+package nz.eloque.foss_wallet.model.field
+
+import androidx.room.Entity
+import nz.eloque.foss_wallet.utils.inIgnoreCase
+import org.json.JSONObject
+
+
+@Entity
+data class PassField(
+    val key: String,
+    val label: String,
+    val value: String,
+) {
+    fun toJson(): JSONObject {
+        return JSONObject().also {
+            it.put("key", key)
+            it.put("label", label)
+            it.put("value", value)
+        }
+    }
+
+    fun contains(query: String): Boolean {
+        return query inIgnoreCase this.label || query inIgnoreCase this.value
+    }
+
+    companion object {
+        fun fromJson(json: JSONObject): PassField {
+            return PassField(json.getString("key"), json.getString("label"), json.getString("value"))
+        }
+    }
+}
